pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.115.8",
    "flask-caching>=2.3.1",
    "flask==2.3.3",
    "flask-sqlalchemy>=3.1.1",
    "httpx>=0.28.1",
    "psycopg2-binary==2.9.10",
    "pydantic==2.10.6",
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.3",
    "python-dotenv==1.0.1",
    "sqlalchemy==2.0.38",
    "uvicorn==0.34.0",
    "celery>=5.4.0",
    "flask-socketio>=5.5.1",
    "eventlet>=0.39.0",
    "gunicorn>=23.0.0",
    "redis>=5.2.1",
    "dash-core-components>=2.0.0",
    "dash-html-components>=2.0.0",
    "dash==2.9.3",
    "dash-bootstrap-components>=1.7.1",
    "pandas>=2.2.3",
    "plotly>=6.0.0",
    "requests>=2.32.3",
    "langchain>=0.3.19",
    "openai>=1.65.0",
    "numpy>=1.26.4",
    "beautifulsoup4>=4.13.3",
    "google-generativeai>=0.8.4",
    "google-api-python-client>=2.162.0",
    "bokeh>=3.6.3",
    "torch>=2.6.0",
    "transformers>=4.49.0",
    "pymupdf>=1.25.3",
    "pytesseract>=0.3.13",
    "pillow>=11.1.0",
    "fpdf>=1.7.2",
    "pymongo[srv]>=4.11.2",
    "motor>=3.7.0",
    "bson>=0.5.10",
    "pinecone-client==3.0.0",
    "psutil>=7.0.0",
    "werkzeug>=3.1.3",
    "dnspython>=2.7.0",
    "scipy>=1.15.2",
]

[[tool.uv.index]]
explicit = true
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"

[tool.uv.sources]
torch = [{ index = "pytorch-cpu", marker = "platform_system == 'Linux'" }]
torchvision = [{ index = "pytorch-cpu", marker = "platform_system == 'Linux'" }]
